Ori dev

+++ compose-oneliner.sh
@@ -18,7 +18,6 @@ SCRIPT=$(readlink -f "$0")
 # Absolute path to the script directory
 BASEDIR=$(dirname "$SCRIPT")
 HOME_DIR=`eval echo ~$(logname)`
-DOCKER_COMPOSE_DIR=${HOME_DIR}/docker-compose
 COMPOSE_BASH_URL="https://github.com/AnyVisionltd"
 
 
@@ -141,17 +140,18 @@ if [[ -z ${DASHBOARD} ]]; then
     DASHBOARD="false"
 fi
 
-if [[ "$DASHBOARD"=="false"  && ! -z $DASHBOARD_VERSION ]]; then
+if [[ "$DASHBOARD" == "false"  && -n $DASHBOARD_VERSION ]]; then
     echo "--download-dashboard was not spcify ignoring --dashboard-version"
     unset DASHBOARD_VERSION
 fi
+
+DOCKER_COMPOSE_DIR=${HOME_DIR}/${GIT}
 
 if [ -x "$(command -v apt-get)" ]; then
 
 	# install git
 	echo "Installing git"
-	git --version > /dev/null 2>&1
-	if [ $? != 0 ]; then
+	if git --version > /dev/null 2>&1; then
 	    set -e
 	    apt-get -qq update > /dev/null
 	    apt-get -qq install -y --no-install-recommends git curl > /dev/null
@@ -164,7 +164,7 @@ fi
 
 if [[ $TOKEN != "" ]] && [[ $TOKEN == *".json" ]] && [[ -f $TOKEN ]] ;then
     gcr_user="_json_key" 
-    gcr_key="$(cat ${TOKEN} | tr '\n' ' ')"
+    gcr_key="$(< ${TOKEN} tr '\n' ' ')"
 elif  [[ $TOKEN != "" ]] && [[ ! -f $TOKEN ]] && [[ $TOKEN != *".json" ]]; then
     gcr_user="oauth2accesstoken"
     gcr_key=$TOKEN
@@ -172,10 +172,9 @@ fi
 
 COMPOSE_REPO="${COMPOSE_BASH_URL}/${GIT}.git"
 [ -d $DOCKER_COMPOSE_DIR ] || mkdir $DOCKER_COMPOSE_DIR
-[ -d ${DOCKER_COMPOSE_DIR}/${BRANCH} ] && rm -rf ${DOCKER_COMPOSE_DIR}/${BRANCH}
+[ -d ${DOCKER_COMPOSE_DIR}/${BRANCH} ] && rm -rf ${DOCKER_COMPOSE_DIR:?}/${BRANCH:?}
 
-git clone ${COMPOSE_REPO} -b ${BRANCH} ${DOCKER_COMPOSE_DIR}/${BRANCH}
-if [ $? -ne 0 ]; then
+if git clone ${COMPOSE_REPO} -b ${BRANCH} ${DOCKER_COMPOSE_DIR}/${BRANCH}; then
     echo "No such branch try again"
     exit 1
 fi
@@ -210,10 +209,9 @@ set +e
 [ -d /opt/compose-oneliner ] && rm -rf /opt/compose-oneliner
 git clone --recurse-submodules  https://github.com/AnyVisionltd/compose-oneliner.git /opt/compose-oneliner
 pushd /opt/compose-oneliner
-ansible-playbook --become --become-user=root ansible/main.yml -vv
 
 
-if [ $? != 0 ]; then
+if ansible-playbook --become --become-user=root ansible/main.yml -vv; then
     echo "" 
     echo "Installation failed, please contact support." 
     exit 1
